+++ components/videodeliverytabs/create_videoclips.js/videoeditor.js
@@ -2,27 +2,24 @@ import { Fragment } from "react";
 import styles from '../../../styles/videodelivery_tabs.module.css';
 import Link from "next/link";
 function Videoditor() {
-  return (
- 
-  <div className={styles.Videodelivery_addnewassets}>
-                    <h2>Upload subtitle files</h2>
-                   
+    return (
 
-                
-                    <div className={styles.post}>
-                     
-                        <div className={styles.language_select}>
-                            <button className={styles.model_btn}>Python</button>
-                            <button className={styles.model_btn}>Node</button>
-                            <button className={styles.model_btn}>Php</button>
-                            <button className={styles.model_btn}>Go</button>
-                        </div>
-                        <div className={styles.code}>
-
-
-                        </div>
-                        <button className={styles.btn}>Run Request</button>
-                    </div>
-                </div> 
-  )}
-  export default Videoditor;
+        <div className={styles.video_editor_content}>
+            <div className={styles.video_editor_player}></div>
+            <div className={styles.start_end_time}>
+                <div className={styles.start_time}>
+                    <h5>Start time*</h5>
+                    <input placeholder="00:20:30"></input>
+                </div>
+                <div className={styles.end_time}>
+                    <h5>End Time*</h5>
+                    <input placeholder="00:20:40"></input>
+                </div>
+            </div>
+            <div className={styles.create_video_clip_button}>
+                <button>Create Video Clip</button>
+            </div>
+        </div>
+    )
+}
+export default Videoditor;
