+++ styles/videoclips_tabs.module.css
@@ -1,0 +1,71 @@
+
+.wrapper_tabs{
+    width: 100%;
+    float: left;
+    padding-top: 40px;
+}
+.Activities_tabs .wrapper_tabs{
+    float: none;
+    padding-top: 0px;
+}
+.Activities_tabs .bloc_tabs{
+    display: block;
+ 
+    
+}
+.bloc_tabs{
+    display: flex;
+    margin-bottom: 20px;
+    border-bottom: 1px solid #d5dbe2;   
+    background-color: #fff;
+}
+.bloc_tabs span{
+    font-family:Inter-Regular;
+    font-size: 14px;
+}
+.tabs_item{
+  padding: 10px 15px 10px 0px;
+  text-align: center;
+  /* width: 15%; */
+  background: transparent;
+  color: #56606e;
+  cursor: pointer;
+  border: none;
+  box-sizing: content-box;
+  position: relative;
+  outline: none;
+  font-weight: 500;
+}
+.active_tabs{
+    /* background: white; */
+    color: #212b36;
+}  
+.active_tabs span{
+    position: relative;
+}
+.active_tabs span::before{
+    content: "";
+    display: block;
+    position: absolute;
+    bottom: -11px;
+    left: 0;
+    width: 100%;
+    height: 2px;
+    background: #2863eb;
+}
+.content_tabs_area{
+    flex-grow : 1;
+}
+.content_tabs{
+    /* background: white; */
+    /* padding: 20px; */
+    width: 100%;
+    height: 100%;
+    display: none;
+}
+.active_content{
+    display: block;
+}
+.videoclips .Activities_tabs{
+    width: 48%;
+}
